update profile: add change profile data button and cleanup code

+++ src/profile_handlers/main_menu_handler.py
@@ -3,22 +3,25 @@ from telegram.ext import ContextTypes, MessageHandler, filters
 
 from databaseAPI import rep_chess_db
 from util import escape_special_symbols
+from change_profile_handler import change_profile_keyboard
 
 
-profile_inline_keyboard = InlineKeyboardMarkup([
-    [
-        InlineKeyboardButton("📝  Ник", callback_data="profile_nickname"),
-        InlineKeyboardButton("📝  Имя", callback_data="profile_name"),
-    ],
-    [
-        InlineKeyboardButton("📝  Фамилия", callback_data="profile_surname"),
-    ],
-    [
-        InlineKeyboardButton("♞  lichess", callback_data="profile_lichess_rating"),
-        InlineKeyboardButton("♟️  chess.com", callback_data="profile_chesscom_rating")
-    ],
+profile_keyboard = InlineKeyboardMarkup([
+    [InlineKeyboardButton("📝  Поменять данные", callback_data="change_profile_data")],
     [InlineKeyboardButton("<< Назад", callback_data="go_main_menu")],
 ])
+
+
+async def change_profile_data(update: Update, context: ContextTypes.DEFAULT_TYPE):
+    query = update.callback_query
+    await query.answer()
+
+    context.bot.send_message(
+        update.effective_chat.id,
+        "_Вы можете поменять свои данные:_",
+        parse_mode="markdown",
+        reply_markup=change_profile_keyboard
+    )
 
 
 def construct_profile_message(user_db_data: dict) -> str:
@@ -34,6 +37,8 @@ def construct_profile_message(user_db_data: dict) -> str:
     profile_str +=  f" ├ Имя:  `{escape_special_symbols(user_db_data['name'])}`\n"
     if user_db_data['surname']:
         profile_str += f" ├ Фамилия:  `{escape_special_symbols(user_db_data['surname'])}`\n"
+    if user_db_data['age']:
+        profile_str += f" ├ Возраст:  `{user_db_data['age']}`\n"
     profile_str = change_last_symbol(profile_str, "├", "└")
     profile_str += f"\n📊 *_Статистика:_*\n"
     profile_str += f" ├ Rep рейтинг:  `{user_db_data['rep_rating']}`\n"
@@ -50,22 +55,24 @@ async def main_menu_handler(update: Update, context: ContextTypes.DEFAULT_TYPE):
     user_db_data = rep_chess_db.get_user_on_telegram_id(telegram_id)
     rep_chess_db.update_user_last_contact(telegram_id)
 
-    # Add user data in cache to not make query to database every time
-    if "user_db_data" not in context.user_data:
-        context.user_data["user_db_data"] = user_db_data
+    # Add user data in cache to not make database query every time
+    context.user_data["user_db_data"] = user_db_data
 
     # Delete saved state because here we already don't expect that useful user message will come.
     context.user_data["text_state"] = None
 
     # Delete useless messages about correcting some data
-    if "messages_to_delete" in context.user_data:
+    if "messages_to_delete" in context.user_data and context.user_data["messages_to_delete"]:
         await context.bot.delete_messages(update.effective_chat.id, context.user_data["messages_to_delete"])
     context.user_data["messages_to_delete"] = []
 
     profile_str = construct_profile_message(user_db_data)
-    message = await update.message.reply_text(profile_str, parse_mode="MarkdownV2", disable_web_page_preview=True)
-    context.user_data["messages_to_delete"].append(message.message_id)
-    message = await update.message.reply_text("Можете поменять данные:", reply_markup=profile_inline_keyboard)
+    message = await update.message.reply_text(
+        profile_str,
+        parse_mode="MarkdownV2",
+        disable_web_page_preview=True,
+        reply_markup=profile_keyboard
+    )
     context.user_data["messages_to_delete"].append(message.message_id)
